+++ src/main/java/com/itsight/domain/jsonb/PostDetalle.java
@@ -12,7 +12,7 @@ import java.util.Date;
 @Data
 public class PostDetalle implements Serializable {
 
-    private int cliId;
+    private Long cliId;
 
     private String nomFull;
 
@@ -31,7 +31,7 @@ public class PostDetalle implements Serializable {
     public PostDetalle() {
     }
 
-    public PostDetalle(int cliId, String nomFull, boolean flgLiked, boolean flgFav, Date feCreac, Date feModif) {
+    public PostDetalle(Long cliId, String nomFull, boolean flgLiked, boolean flgFav, Date feCreac, Date feModif) {
         this.cliId = cliId;
         this.nomFull = nomFull;
         this.flgLiked = flgLiked;
