+++ src/main/java/com/itsight/repository/DiaPlantillaRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DiaPlantillaRepository extends JpaRepository<DiaPlantilla, Integer> {
+public interface DiaPlantillaRepository extends JpaRepository<DiaPlantilla, Long> {
 
     @Override
     @EntityGraph("diaPlantilla")
@@ -19,38 +19,38 @@ public interface DiaPlantillaRepository extends JpaRepository<DiaPlantilla, Inte
 
     @Override
     @EntityGraph("diaPlantilla")
-    DiaPlantilla findOne(Integer id);
+    DiaPlantilla findOne(Long id);
 
     @Query("SELECT D.id FROM DiaPlantilla D WHERE D.semanaPlantilla.id = ?1 ORDER BY 1")
-    List<Integer> findIdBySemanaId(int semanaId);
+    List<Integer> findIdBySemanaId(Long semanaId);
 
     @Modifying
     @Query(value = "UPDATE dia_plantilla SET listas = jsonb_set(listas, CAST(:texto as text[]), to_jsonb(CAST(:nombre as text)), false) WHERE dia_plantilla_id = :id", nativeQuery = true)
-    void updateEspecificaColumnaJsonBGenericoByQueryTextAndId(@Param("id") int id, @Param("nombre") String nombre, @Param(value = "texto") String texto);
+    void updateEspecificaColumnaJsonBGenericoByQueryTextAndId(@Param("id") Long id, @Param("nombre") String nombre, @Param(value = "texto") String texto);
 
     //COALESCE retorna el primer valor not null(en caso no haya ninguna lista se guarde una lista vacia)
     @Modifying
     @Query(value = "UPDATE dia_plantilla SET listas = COALESCE(listas, '[]') || CAST(:lista as jsonb) WHERE dia_plantilla_id = :id", nativeQuery = true)
-    void saveListaDia(@Param("id") int id, @Param("lista") String lista);
+    void saveListaDia(@Param("id") Long id, @Param("lista") String lista);
 
     @Modifying
     @Query(value = "UPDATE dia_plantilla SET listas = jsonb_set(listas, CAST(:texto as text[]), CAST(:elemento as jsonb), true) WHERE dia_plantilla_id = :id", nativeQuery = true)
-    void saveElemento(@Param("id") int id, @Param("texto") String texto, @Param("elemento") String elemento);
+    void saveElemento(@Param("id") Long id, @Param("texto") String texto, @Param("elemento") String elemento);
 
     @Modifying
     @Query(value = "UPDATE dia_plantilla SET listas = jsonb_set(listas, CAST(:texto as text[]), COALESCE(listas->:listaIx->'elementos') || CAST(:elementos as jsonb), true) WHERE dia_plantilla_id = :id", nativeQuery = true)
-    void saveElementos(@Param("id") int id, @Param("listaIx") int listaIx ,@Param("texto") String texto, @Param("elementos") String elementos);
+    void saveElementos(@Param("id") Long id, @Param("listaIx") int listaIx ,@Param("texto") String texto, @Param("elementos") String elementos);
 
     @Modifying
     @Query("UPDATE DiaPlantilla D SET D.flagDescanso = ?2, D.listas = null WHERE D.id = ?1 ")
-    void updateFlagDescanso(int id, boolean flagDescanso);
+    void updateFlagDescanso(Long id, boolean flagDescanso);
 
     @Modifying
     @Query(value = "UPDATE dia_plantilla SET listas = (SELECT listas-?2 FROM dia_plantilla WHERE dia_plantilla_id=?1) WHERE dia_plantilla_id=?1", nativeQuery = true)
-    void deleteListaById(int id, int listaIndice);
+    void deleteListaById(Long id, int listaIndice);
 
     @Modifying
     @Query(value = "UPDATE dia_plantilla SET listas = (SELECT listas #- CAST(?2 as text[]) FROM dia_plantilla WHERE dia_plantilla_id=?1) WHERE dia_plantilla_id=?1", nativeQuery = true)
-    void eliminarElementoById(int id, String texto);
+    void eliminarElementoById(Long id, String texto);
 
 }
