+++ src/main/java/com/itsight/repository/SecurityRoleRepository.java
@@ -1,7 +1,6 @@
 package com.itsight.repository;
 
 import com.itsight.domain.SecurityRole;
-import com.itsight.domain.SecurityUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -10,14 +9,14 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Repository
 @Transactional
-public interface SecurityRoleRepository extends JpaRepository<SecurityRole, Integer> {
+public interface SecurityRoleRepository extends JpaRepository<SecurityRole, Long> {
 
     SecurityRole findBySecurityUserUsername(String username);
 
-    SecurityRole findBySecurityUserIdAndRole(int securityUserId, String role);
+    SecurityRole findBySecurityUserIdAndRole(Long securityUserId, String role);
 
     @Modifying
     @Query("DELETE FROM SecurityRole S WHERE S.id = ?1")
-    void deleteById(int id);
+    void deleteById(Long id);
 
 }
